all bugs are fixed

+++ src/utils/cloudinary.js
@@ -1,32 +1,31 @@
-// Usable code
-// It gets the file in local storage then with help of local path it stores it in servers(cloudinary server) and 
-// deletes the files form temporary storage
-
 import {v2 as cloudinary} from "cloudinary"
+import fs from "fs"
 
-    // Configuration
-    cloudinary.config({ 
-        cloud_name: 'process.env.CLOUDINARY_CLOUD_NAME', 
-        api_key: 'process.env.CLOUDINARY_API_KEYS', 
-        api_secret: '<process.env. CLOUDINARY_API_SECRET>' 
-    });
+const uploadOnCloudinary = async (localFilePath) => {
+    try {
+        if (!localFilePath) return null
 
-    const uploadOnCloudinary = async (localFilePath) => {
-        try {
-            if(!localFilePath) return null;
-            //upload file on cloudinary
-            const response = await cloudinary.uploader.upload( localFilePath, {
-                resource_type : "auto"  //for file type
-            } )
-            //file has been upload successfully
-            console.log("file is uploaded on cloudinary", response.url); //we take the url after upload on cloudinary
-            return response;
+            cloudinary.config({ 
+            cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+            api_key: process.env.CLOUDINARY_API_KEY, 
+            api_secret: process.env.CLOUDINARY_API_SECRET
+            });
+        
+        const response = await cloudinary.uploader.upload(localFilePath, {
+            resource_type: "auto"
+        })
+        
+        // console.log("File uploaded:", response.url);
+        fs.unlinkSync(localFilePath)
+        return response;
 
-        } catch (error) {
-            //remove the localyy saved temporary file as the uploaf operation got failed
+    } catch (error) {
+        console.log("Upload failed:", error.message);
+        if(fs.existsSync(localFilePath)) {
             fs.unlinkSync(localFilePath)
-            return null;
         }
+        return null;
     }
-    
-export {uploadOnCloudinary};
+}
+
+export {uploadOnCloudinary}
